pyproject.toml
[project]
name = "matcha-backend"
version = "0.1.0"
description = "Pragmatic FastAPI MVP backend scaffold for MATCHA."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "pydantic-settings>=2.8.0,<3.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "alembic>=1.13,<2.0",
    "asyncpg>=0.29,<1.0",
    # Auth
    "python-jose[cryptography]>=3.3,<4.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "bcrypt>=4.0,<5.0",
    "python-multipart>=0.0.9,<1.0.0",
    # Cache
    "redis[hiredis]>=5.0,<6.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.23,<1.0",
    "aiosqlite>=0.20,<1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
